Fix comments in all rule of Makefile.example
Without the @ before the comments (which are informative and to be
removed by contestants) it shows the comments when running 'make'. Even
worse is without doing $${CHMOD} it would not evaluate to anything so it
looked bizarre.

Thus the @ was added to the comments like the rest of the file (comments
in rules).