Final edits before IOCCC28 (I hope :-) )
Renamed some FAQs (typos and/or other things).

Fix a number of typos in various files (including markdown format
errors).

Rules and guidelines should be relatively good although as there is a
lot to get through and not much time left maybe I missed something (I
also kind of cheated doing the final checks with the guidelines rather
than going through them all again).

The next/try.sh and next/try.alt.sh scripts have had updates as well and
the next/Makefile.example was modified too.

The index.html page now has an anchor for goals of the contest so that
the goals in the rules could simply link to it in index.html.

Although some text had to be added (and some corrected) a reasonable
amount of text was also removed as it was elsewhere or was extraneous in
some way.