Updated FAQ per recent mkiocccentry issues
The warnings highbit_warning, wordbuf_warning and first_rule_is_all are
no longer used and will be removed for IOCCC29 and beyond. Thus the
tools no longer care about them.

There were other additions to mkiocccentry like the useful -u uuid
option to read in your UUID from a file so you don't have to copy paste
it per submission. Although that would be useful in the guidelines to
do that would require the version update which is problematic in an open
contest. Thus it can be added for IOCCC29 after this one.

IMPORTANT NOTE: you do NOT have to update the tools unless you want to!
The submit server will no longer flag submissions with those booleans
set to true and when the judges run the tools they won't flag them
either. So if you are just fine with what you're doing, and if the
recent fixes do not have an impact on your work you are perfectly fine
using the current versions you have as version requirements are no
longer a single value but a range. Of course you might still wish to
check that you weren't flagged for some other issue but you no longer
have to worry about those three warnings.