Add details on versions
New FAQ on what the minimum versions are (this should be updated every
contest or updated to refer to the result of the issue I opened at the
other repo - post IOCCC28) for each tool.

Added details about the new (soon to be released) version of the tools
that now search under $PATH.

This commit updated the FAQ, the guidelines and the quick start guide.

As a reminder you do NOT need to use the upcoming release of the tools
and not doing so will not invalidate any of your submissions as it is
based on a minimum version, not an exact version.