improve use of chkentry by cvt-submission.sh
Modified `bin/cvt-submission.sh` to use `chkentry(1)` on
submissions that are about to be converted into an entry,
by using `-P` to ignore permissions.  We also ignore any
`.DS_Store` files found as they will not be committed to
the repo (because of the top level `.gitignore` file).

Changed `bin/cvt-submission.sh` VERSION from "2.2.0 2025-06-29"
to "2.2.1 2025-06-29".