Add FAQ about enabling mastodon push notifications
It is not clear, as the FAQ shows for now, whether this works with just
desktop/computers or not - it is only possible to enable it on computers
(or if it's possible in mobile devices I did not see it though I only
checked with an iPhone).

This FAQ should probably be linked to in the guidelines later on as
right now it says that one should reload the feed as you do not get
notifications unless you are mentioned (assuming you have push
notifications enabled). As one of the pending changes post winners being
announced (02 August 2025 at 11 Pacific) is to clean up the guidelines,
rules and maybe the FAQ (that was rushed towards the end so as to not
delay IOCCC28 even longer) I have not done this.

Rebuilt faq.html. New image under png/.

FAQ version: 28.2.21 2025-07-31.