fix 2024/anders swap 2024/straadt
We swapped `2024/anders` and `2024/straadt` by accident.
Sorry (tm Canada 🇨🇦) about that!

We **SINCERELY APOLOGIZE** to both **Anders Kaare Straadt**
and **Sebastian Macke** for the mix up!

We normalized some of the `.gitignore` files for some
of the 2024 winning entries.

Rebuilt 2024 tar files.

Updated `sitemap.xml`.

Ran `make update` to perform the above.