Typo fixes in 2024/README.md
Also rebuilt 2024/guidelines.html, 2024/rules.html (obviously
2024/index.html was also rebuilt).

The other ones were rebuilt because the markdown had 2025 (from 2024)
and it was not rebuilt. I noticed a lot of files have this problem (in
copyright). May I suggest sgit(1) to save you time? :-) (I would do it
but I don't want to cause a problem with any pull requests that might
be opened.