Add new file to 2024/ferguson1
The gotos.txt file (manifest updated) shows the line changes in a single
run of the program.

Capitalised some text.

Changed width and height attributes in images - they were too hard to
see otherwise (for those of us with poor vision it's still hard on a
mobile device but it was almost impossible to see on the computer for
me - now it scales better, for some definitions of 'better').

Rebuilt html files.