Manifest fixes
2012/omoikane had some updates but the manifest was not properly updated
(so it would seem - it seems right but I cannot figure out how it was
not as obviously the CSV file was correct).

For some reason 2020/yang/index.html navigation linked to 2024/anders
which of course is a 404 (I seem to recall a mix-up that was fixed but
the html files were not rebuilt).

The 2020/yang/.entry.json file had the wrong manifest version as well.