Fix html links in 2024/
Due to a manifest mix-up (which had some html files marked as markdown
and as display_via_github as true) the respective entries' index.html
linked to html files on GitHub which do not render right (and is the
wrong place in any case).

This means that the index.html files and the .entry.json files were
updated. This in particular goes for:

    2024/burton
    2024/ferguson1
    2024/howe

Thanks to @SirWumpus for reporting this!