Fix 2024/weaver/.entry.json, rebuilt index.html
The entry_text for 'ref.mp3' had commas. This is not allowed as the way
the site script (in particular bin/entry2csv.sh) works is it creates a
CSV file which when imported into a spreadsheet application (like
Numbers in macOS) it adds additional columns. Also, because a space is
after the comma, it triggers another error in the bin/csv2entry.sh
script.

What this error meant is that some of the text was not even in the
index.html.

This has all been corrected.