Add fun file to 2024/ferguson1
The file prog.gotos.c has the lines merged so one can use their syntax
highlighter (or grep, see below) to see all the gotos at once (depending
on screen size, perhaps).

For grep try:

    grep -E 'goto|hell|heaven|666' prog.gotos.c

if you have colours enabled. The other patterns is because of alignment
that was maintained in the program (one could do this without 'goto' of
course).