Various fixes, improvements in 2024/ferguson1
The alt code was an old version that should not have even been there.
The extra event was something that was meant to be somewhere else (like
it is in the prog.c) and the event (even 23) is something else.

I put in the data files another joke with the intention of using that
for the event until I remembered the event was wrong.

The alt code has an improvement: it now is possible to specify your
hp/stamina and max number of turns.

In prog.c remove an unused variable.

Updated Judges' remarks to better reflect what happens (in particular
improve the commits from yesterday - correcting it but keeping the same
flow of the judges).

Updated deobfuscation.md (rebuilt deobfuscation.html).

Rebuilt index.html.