Update bugs.html for 2024/ferguson2 - INABIAF
This is related to the space / word boundary issue (based on how the
language worked) that does not have a good solution. As I think I might
have written in the remarks I tried different ways but each way caused
other problems, problems that would make Navajo -> English not work even
when it should otherwise. Although the index.html already showed some
examples of this it seems good to include some in the bugs.html and add
the INABIAF tag to the index.html.