Add tips to 2020/ferguson1/chocolate-cake.html
At the suggestion of Landon I have added some tips to both icing and the
cake. It might be possible for more to be added later on but hopefully
these are helpful to those who wish to try it. Something I will try
doing is adding sequential pictures of the icing but that involves
remembering it as we can't make it here until the winter (well we can
but it's a lot harder as even the winter can be fairly hot here,
exceeding 35 Celsius).

And to those who (presumably many) do wonder why I submitted this back
then when it has nothing to do with the contest: I really haven't the
foggiest but since this was requested I am doing it.