Add another event to 2024/ferguson1 alt code
This was somehow not added during development but was supposed to be. It
is not going in the main code because that would ruin the amount of
gotos (152 for full version, 134 for practical version) and because the
point of the alt code was to have more events anyway.

Tested this new event. Works fine.