Fix CSILENCE= in next/Makefile.example
One of the warnings disabled (triggered in macOS) that was already added
seemed to be missing.

I also disabled the silly warning triggered by clang with -Weverything (I
think -Weverything) -Wunsafe-buffer-usage as that is against basic C (it
will not even let you iterate through argv without whining about it
being 'unsafe').

The example in the Makefile given is -Wno-int-conversion instead.