Add -Wno-overriding-deployment-version
A bug in macOS 26 (is anyone surprised any more?) ends up with warning:

    clang: warning: overriding deployment version from '16.0' to '26.0' [-Woverriding-deployment-version]

depending on invocation. Particularly with at least some versions of
gcc, even though in this case it was actually gcc and not the system gcc
that is really clang in disguise.

Because of this -Wno-overriding-deployment-version was added to
next/Makefile.example.