Update FAQ, rules/guidelines etc. for #208
This commit also removed some of the 'new marks' in the guidelines and
rules (as they're no longer new) and some that were in the FAQ (which
only happened when they were changed over to the FAQ).

Removed problematic FAQs (per discussion). This required renumbering.

Updated references of chkentry to chksubmit (only a few references to
chkentry remain now - advising contestants to not use it but simply
stating it is for the judges and it is what chksubmit uses).

Added note in guidelines about gotos becoming 'old' (as Landon told me
it is something Leo wanted in the next guidelines thanks to me). Just so
Landon does not forget my C sins. :-)

I have regenerated the html files but the guidelines.html and rules.html
files have not been link-checked as these files will be modified by the
judges a fair amount so there is no point yet. I believe I got
references to nuked FAQs out but it's possible some were missed.

The quick-start.md and next/submit.md (and their html files) were
updated as well to refer to chksubmit instead of chkentry.

It is quite possible some things were missed but this should help move
things closer to getting ready for IOCCC29 in December.