Correct comments wrt CC in example Makefile
A long while back I improved the Makefiles to allow for compiler
specific flags so that they don't have to be only "clang" or "gcc". This
is because in some systems gcc is actually clang (macOS does this). Also
if someone wants to use a different version or path then they can do so
with the right flags enabled. Before that change it would not work
right.

The comments suggested these flags will only be set when CC is exactly
"clang" or "gcc". This has been corrected.