Fix next/try.alt.sh, update try script versions
The try.alt.sh used 'make all' and it also called './prog' and not
'./prog.alt'.

The version for both try.sh and try.alt.sh were updated. Presumably
before the contest is opened this will change again (or not due to this)
but since I edited the try.alt.sh and updated the version it seems
fitting to update the try.sh version too.