Fix bin/pandoc-wrapper to use verge(1) - #237
The script needs to use verge(1) (see
https://github.com/xexyl/jparse/blob/master/verge.c) from the jparse
repo (see https://github.com/xexyl/jparse/).

NOTE: it is NOT recommended you install verge(1) from the jparse repo
for IOCCC related things. Instead run (the first line only if you don't
already have the mkiocccentry repo cloned and updated; if you have it
you might wish to pull the current release):

    git clone https://github.com/ioccc-src/mkiocccentry.git"
    cd mkiocccentry && make clobber all
    cd jparse && sudo make install clobber

This is because on a rare occasion the jparse repo can be ahead of the
mkiocccentry's jparse/ directory.

Tested with 'make markdown' (this is enough because it uses the
pandoc-wrapper.sh and it's also what triggered the bug for @SirWumpus in
NetBSD - which does not support sort -V).

In theory the script should work under NetBSD now.