Fixes and minor improvements in rules
This includes typos (including command lines, formatting and other
things), renaming some rules to be a bit more intuitive and other things
I noticed.

Please note that I have not yet checked all links: not in other files or
in this file. This will be done after a break, and it likely will
involve updates to the guidelines and possibly the FAQ too. In other
words this is a work in progress that should be completed later today.