Various fixes and updates
The links to rules should all be good. This includes in news.html,
quick-start.html, next/submit.html, 2024/index.html (it had been
modified to refer to next/rules.html instead of 2024/rules.html) as well
as the guidelines and rules themselves (under next/).

Various typo fixes. I'm afraid there are likely others but I must let
others deal with that as I am out of time.

Removed FAQ about mastodon push notifications. When I added the FAQ
about push notifications for mastodon I was incorrect. The web browser
interface suggests it but I never got a push notification even before
Apple broke them. There seems to be no way at this time to get them. I
note that there was a screenshot in the FAQ but this file still is in
the repo possibly.

I regenerated the html files but it's possible I forgot to do one at the
last minute.