Restore guidelines
Somehow the guidelines got clobbered during an edit. It is not clear to
me if every link is okay in this but from a quick look at it it some of
the links if not all were certainly updated. It will require another
pass in the browser but this is an emergency commit.