fix guidelines, remove guideline status, update sitemap
This is a checkpoint correction for issue #244 ("Restore guidelines")
which in turn was a correction for issue #243 ("Fixes and minor
improvements in rules").

Remove code that put the contest status at the top
of the guidelines page.

Change `bin/gen-status.sh` VERSION to "2.1.0 2025-11-29",

Fix `next/guidelines.md` so that `next/guidelines.html`
is rendered correctly.

Update sitemap and status (due to recent news mod).

Ran `make quick_www` to perform the above

TODO: Complete the correction for issue #244 ("Restore guidelines").

TODO: Continue to revise and improve the IOCCC guidelines as needed.

TODO: Fix typos/nits in the `next/guidelines.md` if needed.

TODO: Fix typos/nits in the `next/rules.md` if needed.