checkpoint guideline fixes
Attempt to fix HTML5 errors.

Remove Makefile section from Guidelines as
this is in the FAQ.

Added links to Makefile FAQ in Guidelines.

Ran `make quick_www` to perform the above

TODO: Complete the correction for issue #244 ("Restore guidelines").

TODO: Continue to revise and improve the IOCCC guidelines as needed.

TODO: Fix typos/nits in the `next/guidelines.md` if needed.

TODO: Fix typos/nits in the `next/rules.md` if needed.