Update FAQ and fix guidelines
The FAQ has a new item that is referred to in the guidelines updates.

The guidelines has some link fixes and other minor changes.

Some fun was corrected in the guidelines as well but kept a fair bit
shorter than before (i.e. shorter than IOCCC28). In particular because
some code from the mkiocccentry (from some contributors) is no longer
extant, and because said code was not even a complete program, only
refer to two programs that were strictly written by someone other than
the judges.  Those paragraphs are only one sentence (they used to be
more than one) and the rest of the mkiocccentry repos is not mentioned
(this was the longest one and it went into other things that while fun
were not relevant). This keeps it light, fun, very short and more
importantly correct both to the rule about submissions must be a
COMPLETE C program and other rules as well.

Some typo fixes were made.

I believe this commit resolves the problems from yesterday although I
did not go through the file extremely carefully as I know @SirWumpus
also wants to go through it and so does Landon, so I kind of rushed
through it.