Final fixes for IOCCC guidelines/rules
Fixed broken links.

Restored text that was accidentally removed during the emergency
recovery of guidelines some days back.

Moved some text that was in the wrong place.

The formatting of the markdown files was changed for shorter lines (for
those viewing the markdown in a viewer rather than a web browser).

Added a <div id> to 2024/index.html so the guidelines (or maybe rules)
could link to the important note from the judges in IOCCC28.

Restore rewording of rule 2b done by @SirWumpus.