Fixes in guidelines
This includes a number of blocks of text in the wrong place.

Some typos were fixed but just like with the rules a very thorough job
was not done and there are deliberate typos in the guidelines as well.
It should be a bit clearer now, except perhaps the guidelines that are
deliberately confusing (or those I missed).